refactor: rename utils.ts to date-utils.ts
