+++ src/components/section/EventInfoHeader.tsx
@@ -16,7 +16,7 @@ import {
   formatOccurring,
   formatTimespan,
   formatTimeZoneTag,
-} from "../../utils/utils";
+} from "../../utils/date-utils";
 import { IoEarthSharp } from "react-icons/io5";
 import { EventTypeTag } from "../ui/Tag";
 import { DeleteDialog, EditDialog, LeaveDialog } from "./Dialog";
