+++ src/components/section/NewEventCard.tsx
@@ -2,7 +2,11 @@ import type { ListboxOption } from "../ui/Input";
 import { RoundedListbox, Selector } from "../ui/Input";
 import { useRouter } from "next/router";
 import React, { useState } from "react";
-import { getCurrentTimeZoneTag, makeTime, moveTime } from "../../utils/utils";
+import {
+  getCurrentTimeZoneTag,
+  makeTime,
+  moveTime,
+} from "../../utils/date-utils";
 import { RoundedTimezoneInput } from "../ui/TimezoneInput";
 import { DateSelect } from "../ui/DateSelect";
 import { TimespanSelector } from "../ui/TimeSelector";
